requirements.txt
black==24.2.0
click==8.1.7
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
